renamed the files to use the plural
